Fix loss of precision in tone portamento
Our periods are defined as doubles, but when performing a tone portamento
the target portamento was stored as an integer. Change that to double so
we can slide to correct periods.

Reported by Misty De Meo, check jazz jackrabbit 2 - prerelease - medivo.it
order 12, channel 10.

+++ src/player.h
@@ -137,7 +137,7 @@ struct channel_data {
 		int slide;	/* Frequency slide value */
 		int fslide;	/* Fine frequency slide value */
 		int memory;	/* Portamento effect memory */
-		int s_end;	/* Target period for tone portamento */
+		double s_end;	/* Target period for tone portamento */
 		int s_sgn;	/* Tone portamento up/down switch */
 		int s_val;	/* Delta for tone portamento */
 		int s_memory;	/* Tone portamento effect memory */
